websocket/requirements.txt
fastapi
uvicorn[standard]
redis
python-dotenv